Acerta deleção da Tecnologia
Acerta o seed tb

+++ database/migrations/2017_04_17_174059_create_responsavels_table.php
@@ -16,7 +16,7 @@ class CreateResponsavelsTable extends Migration
         Schema::create('responsavels', function (Blueprint $table) {
             $table->increments('id');
             $table->string('nome');
-            $table->integer('telefone');
+            $table->bigInteger('telefone');
             $table->string('email');
             $table->integer('tecnologia_id')->unsigned()->nullable();
             $table->foreign('tecnologia_id')
